summary table refactoring
